[WIP] 壁周りのベクトルを順方向に修正

+++ include/render.h
@@ -20,7 +20,7 @@ int draw_vertical_line(
 	size_t vline_len
 );
 
-uint32_t *get_vertical_arr_n(void *img_arr, size_t index, t_vec_i32x4 size, int out_size);
+uint32_t *get_vertical_arr_n(void *img_arr, int index, t_vec_i32x4 size, int out_size);
 
 int render_wall_to_screen(
 	t_data *data,
